Feat: refresh 재발급 로직 수정 및 물건 조회 시 등록자 닉네임 반환

+++ src/main/java/Auction_shop/auction/security/auth/service/AuthService.java
@@ -86,9 +86,13 @@ public class AuthService {
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
         }
 
-        refreshTokenRepository.delete(refreshToken);
+        MemberResponseDto memberResponseDto = memberService.getMemberByRefreshToken(refreshToken);
+        if (memberResponseDto == null) {
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+            return null;
+        }
 
-        MemberResponseDto memberResponseDto = jwtUtil.getMemberResponseDto(refreshToken);
+        refreshTokenRepository.delete(refreshToken);
 
         String accessToken = jwtUtil.createJwt("access", memberResponseDto, accessTokenExpiredMs);
         refreshToken = jwtUtil.createJwt("refresh", memberResponseDto, refreshTokenExpiredMs);
